@@ -1,13 +1,8 @@
 ext {
-    compileSdkVersion = 23
-    buildToolsVersion = "25.0.0"
+    compileSdkVersion = 28
     minSdkVersion = 9
     appMinSdkVersion = 14
-    targetSdkVersion = 23
-
-    libSupportAppcompatV7 = "com.android.support:appcompat-v7:${buildToolsVersion}"
-    libRecycleView = "com.android.support:recyclerview-v7:${buildToolsVersion}"
-
+    targetSdkVersion = 28
 }
 
 buildscript {
@@ -16,7 +11,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.1'
+        classpath 'com.android.tools.build:gradle:3.5.3'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
