@@ -1,8 +1,9 @@
 package com.mayousheng.www.jsondecodepojo.base;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.mayousheng.www.initview.ViewUtils;
 
