@@ -1,7 +1,7 @@
 package com.mayousheng.www.jsondecodepojo.listener;
 
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 public abstract class OnLoadMoreListener extends
         RecyclerView.OnScrollListener {
