@@ -1,11 +1,12 @@
 package com.mayousheng.www.jsondecodepojo.base;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.mayousheng.www.initview.ViewUtils;
 
