@@ -1,8 +1,9 @@
 package com.mayousheng.www.jsondecodepojo.base;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.mayousheng.www.initview.ViewUtils;
 
