@@ -9,9 +9,10 @@ import com.mayousheng.www.jsondecodepojo.base.BaseNewsHolder;
 import com.mayousheng.www.jsondecodepojo.common.StaticParam;
 import com.mayousheng.www.initview.ViewDesc;
 import com.mayousheng.www.jsondecodepojo.pojo.BSBDJPhotoResponse;
-import com.mayousheng.www.jsondecodepojo.utils.ShowImageUtils;
 
 import java.lang.ref.WeakReference;
+
+import www.mayousheng.com.showimgutils.ShowImageUtils;
 
 /**
  * Created by ma kai on 2017/10/5.
