@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mayousheng.www.jsondecodepojo.utils.cache;
+package www.mayousheng.com.showimgutils.cache;
 
 import java.io.ByteArrayOutputStream;
 import java.io.Closeable;
@@ -193,4 +193,3 @@ class StrictLineReader implements Closeable {
     end = result;
   }
 }
-
